Rename functionArgumentsSearch to align with styleSearch
